Fixed SQL comments

+++ db/derby/create-db-position.sql
@@ -3,7 +3,7 @@
 -- requires create-db-common.sql
   
 create table pos_position (
-    id int8 not null,
+    id bigint not null,
     identifier varchar(255) not null,
     startDate date,
     endDate date,
@@ -15,10 +15,10 @@ create table pos_position (
 );
 
 create table pos_domainspecificidentifierassociation (
-    id int8 not null,
+    id bigint not null,
     startDate date,
     endDate date,
-    position_id int8 not null,
+    position_id bigint not null,
     domain varchar(255) not null,
     identifier varchar(255) not null,
     primary key (id),
@@ -27,20 +27,20 @@ create table pos_domainspecificidentifierassociation (
 );
 
 create table pos_portfolionode (
-    id int8 not null,
+    id bigint not null,
     identifier varchar(255) not null,
     startDate date,
     endDate date,
     name varchar(255),
-    ancestor_id int8,
+    ancestor_id bigint,
     primary key (id),
     constraint pos_fk_portfolionode2portfolionode foreign key (ancestor_id) references pos_portfolionode (id),
     unique (identifier, startDate, endDate)
 );
 
 create table pos_nodehierarchy (
-    ancestor_id int8 not null,
-    descendant_id int8 not null,
+    ancestor_id bigint not null,
+    descendant_id bigint not null,
     primary key (ancestor_id, descendant_id),
     constraint pos_fk_nodehierarchy2portfolionode1 foreign key (ancestor_id) references pos_portfolionode (id),
     constraint pos_fk_nodehierarchy2portfolionode2 foreign key (descendant_id) references pos_portfolionode (id)
@@ -48,8 +48,8 @@ create table pos_nodehierarchy (
 create index pos_ix_nodehierarchy on pos_nodehierarchy (descendant_id);
 
 create table pos_nodeinclusion (
-    position_id int8 not null,
-    node_id int8 not null,
+    position_id bigint not null,
+    node_id bigint not null,
     primary key (position_id, node_id),
     constraint fk_nodeinclusion2position foreign key (position_id) references pos_position (id),
     constraint fk_nodeinclusion2portfolionode foreign key (node_id) references pos_portfolionode (id)
@@ -57,12 +57,12 @@ create table pos_nodeinclusion (
 create index pos_ix_nodeinclusion on pos_nodeinclusion (node_id);
 
 create table pos_portfolio (
-    id int8 not null,
+    id bigint not null,
     identifier varchar(255) not null,
     startDate date,
     endDate date,
     name varchar(255) not null,
-    root_id int8 not null,
+    root_id bigint not null,
     primary key (id),
     constraint fk_portfolio2portfolionode foreign key (root_id) references pos_portfolionode (id),
     unique (identifier, startDate, endDate)
